requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-huggingface
langchain-ollama
faiss-cpu
pypdf
pillow
pytesseract
sentence-transformers
numpy
opencv-python-headless